Add files via upload
This project is a single-page application for an e-commerce website built using React.js. The main features of the project include integration with a fake API, as well as the implementation of various components such as Home, Product, Single Product, About, Contact, Navbar, and Footer using React.js and Bootstrap. Redux library is used for state management